Move SharedError to storage/error.rs
